WEB_QR_GEN/requirements.txt
qrcode[pil]>=7.4.2
Pillow>=10.0.0

